[feat/#88] 타입 위치 수정
