+++ app/src/main/java/com/poti/android/presentation/history/participant/model/Contract.kt
@@ -10,20 +10,6 @@ data class ParticipantDetailUiState(
     val overlayState: ParticipantDetailOverlayState = ParticipantDetailOverlayState.None,
 ) : UiState
 
-sealed interface ParticipantDetailOverlayState {
-    data object None : ParticipantDetailOverlayState
-
-    data object DepositBottomSheet : ParticipantDetailOverlayState
-
-    data object DeliveryConfirmModal : ParticipantDetailOverlayState
-
-    data class DeliveryReviewModal(
-        val recruiterName: String,
-        val recruiterProfileUrl: String,
-        val partnerRating: String,
-    ) : ParticipantDetailOverlayState
-}
-
 sealed interface ParticipantDetailUiIntent : UiIntent {
     data class LoadDetail(val recruitId: Long) : ParticipantDetailUiIntent
